ctb: move vm.broadcast into createDeterministic (#14243)
* ctb: Use startbroadcast/stopbroadcast with createDeterministic

This backports https://github.com/ethereum-optimism/optimism/pull/14241 onto develop. While deploying the interop contracts we discovered a bug where nonces would incorrectly increase following calls to `createDeterministic`. This is because `vm.broadcast` will broadcast the _next_ call, but `createDeterministic` sometimes doesn't emit a call at all. This means that (in the case of the Sueprchain deployment) the calls to the input contract's `set` method would be called instead. This in turn caused nonces to increase incorrectly.

To fix this issue, this PR updates the Solidity implementation to use `vm.startBroadcast()` and `vm.stopBroadcast()`. This avoids the issue altogether. To prevent further incorrect usages of `vm.broadcast` I also opened https://github.com/ethereum-optimism/optimism/pull/14242 which adds a Semgrep rule to check for this case.

* move vm.broadcast into createDeterministic