op-program: Fix regeneration of optimistic blocks when they have been pruned (#15741)
* op-program: Fix regeneration of optimistic blocks when they have been pruned.

Update action tests to create a verifier node to run FPP against that only contains canonical blocks so replaced optimistic blocks are not available without regeneration.

* op-program: Update tests to expect fetching output roots and block data.

* op-program: Update baseline to include new call stacks into rawdb.NewStateFreezer

Only the in-memory version is ever actually used but the analyzer can't detect that so reports issues with the alternate path that goes through newResettableFreezer.

* Add comment.

* Improve comment.