op-deployer: Remove tag validation in locator unmarshal (#16181)
Previously, the unmarshaler for the artifacts locator was checking if artifacts locator tags were valid (e.g., a defined tag on the monorepo) in addition to checking to see if they were well formatted. Now that OP Deployer supports one contracts version per binary version, this was causing old state files to fail to unmarshal.