op-e2e/actions: prefer miner.IncludeTx to RPC tx submission (#16255)
This test is flaking around this line because the L1 miner txpool sometimes rejects the transaction citing prague not being active. This should be impossible, because we already performed a synchronous check that prague is active. That check however, does not use the miner's RPC but uses a shortcut through the instrumented op-e2e code. This change uses a similar shortcut to include the desired transaction, which should hopefully avoid any data races and restore consistency.

In any case we are not trying to test L1 behavior here so we should just do whatever is necessary to get that transaction confirmed reliably.