chore(devstack): Don't fail if no conductors are present (#16776)
## Overview

Updates the behavior of `l2_network.go`'s `ExtensibleL2Network` implementation to not fail tests if there are no conductors present in the query to fetch the network's conductors. This function is useful for tests that want to gate on configurations that have conductors present. As-is, it's not possible to do so due to the test failing on query.